fix: resolve 120+ audit findings across codebase
CRITICAL fixes:
- jsprav.py: fix multi-category declared_total using cumulative count
- jsprav.py: handle string telephone/sameAs in JSON-LD parsing
- jsprav.py: guard against None script.string before json.loads
- phone_cluster.py: normalize phones before clustering (dedup was broken)
- checkpoint.py: fix FK ordering in clear_city (Raw before Company)

HIGH fixes:
- _playwright.py: safe resource cleanup in finally block
- tech_extractor.py: case-insensitive CMS detection
- validator.py: rename ALLOWED_HOSTS to INTERNAL_HOSTS, handle IPv6-mapped IPv4
- csv.py/markdown.py: guard None crm_score in sort, safe key access
- markdown.py: complete markdown escaping (* _ # ` ~)
- csv.py: sanitize_filename for city in export paths
- enrichment_phase.py: fix getattr fallback chain
- base.py: initialize last_error in __init__
- merger.py: normalize phones during merge, validate cities
- tg_finder.py: remove invalid phone-as-telegram-username

MEDIUM fixes:
- database.py: standardize timezone, config error handling, remove deprecated
  PipelineRunRow and merged_into column, add __repr__, pool_pre_ping
- cli.py: error handling for config load, validate city entries, fix int(revision)
- scrapers: remove redundant wait_for_load_state, yell extracts all phones,
  fix CSS selector escaping, move imports to module level
- regions.py: thread-safe cache with double-checked locking, YAML validation
- category_finder.py: case-insensitive subdomain regex, fix type hints
- pipeline: centralize STANDARD_SOURCES, firecrawl dead code removal,
  URL validation before subprocess, batch commits in enrichment
- network_detector.py: merge triple iteration into single pass
- name_matcher.py: add cross-prefix secondary pass for small blocks
- messenger_scanner.py: add adaptive_delay between requests
- requirements.txt: remove unused fake-useragent2 and python-dotenv

LOW+TRIVIAL:
- utils.py: preserve email order via dict.fromkeys, fix regex
- status.py: handle unknown log levels
- dedup: type annotations, __all__ exports
- tests: fix mutable dict in lambda, update for removed request_delay
- Add __version__ to package

+++ granite/enrichers/classifier.py
@@ -34,7 +34,7 @@ class Classifier:
             score += self.weights.get("has_telegram", 0)
             
             tg_trust = company.get("tg_trust", {})
-            score += (tg_trust.get("trust_score", 0) * self.weights.get("tg_trust_multiplier", 0))
+            score += int(tg_trust.get("trust_score", 0) * self.weights.get("tg_trust_multiplier", 0))
             
         if messengers.get("whatsapp"):
             score += self.weights.get("has_whatsapp", 0)
